+++ sendbird_platform_sdk/model/list_all_emojis_and_emoji_categories_response_emoji_categories_inner_emojis_inner.py
@@ -82,7 +82,7 @@ class ListAllEmojisAndEmojiCategoriesResponseEmojiCategoriesInnerEmojisInner(Mod
                 and the value is attribute type.
         """
         return {
-            'id': (float,),  # noqa: E501
+            'id': (int,),  # noqa: E501
             'key': (str,),  # noqa: E501
             'url': (str,),  # noqa: E501
         }
@@ -139,7 +139,7 @@ class ListAllEmojisAndEmojiCategoriesResponseEmojiCategoriesInnerEmojisInner(Mod
                                 Animal class but this time we won't travel
                                 through its discriminator because we passed in
                                 _visited_composed_classes = (Animal,)
-            id (float): [optional]  # noqa: E501
+            id (int): [optional]  # noqa: E501
             key (str): [optional]  # noqa: E501
             url (str): [optional]  # noqa: E501
         """
@@ -227,7 +227,7 @@ class ListAllEmojisAndEmojiCategoriesResponseEmojiCategoriesInnerEmojisInner(Mod
                                 Animal class but this time we won't travel
                                 through its discriminator because we passed in
                                 _visited_composed_classes = (Animal,)
-            id (float): [optional]  # noqa: E501
+            id (int): [optional]  # noqa: E501
             key (str): [optional]  # noqa: E501
             url (str): [optional]  # noqa: E501
         """
